Fix downloading generated files
Passing the return value of ProjectIndexView unconditionally to
render_to_response() does not work anymore, so instead move it to
ProjectIndexView to be called conditionally as appropriate.

There are still some issue with the download, Firefox will try to
install the langpack instead of saving it and "Save Link As" will
suggest a misleading "Download langpack.htm" file name.